backend/app/requirements.txt
fastapi>=0.110.0
uvicorn>=0.23.0
pydantic>=2.0
pandas>=2.0
numpy>=1.23
python-dotenv>=1.0
requests>=2.31
